iteration 2
Keeping the feedback in mind, this iteration focuses mainly on challenge. There is now a timer, making high scores possible, and also added some randomness to decaying bottles. There is now also a menu and a reset button for convenience.
A challenge mode is added for a slightly different experience.

(previous push I left out the controls so I am just going to type it here:

r: reset
click with mouse to interact.
click to bottles to merge and gain points. the more decayed the higher the score.)