iteration 3
This iteration I mainly focused on UI/UX.
I noticed from the feedback that a highscore was highly sought after, this in combination with a button on the screen for going back instead of pressing a key to make it more intuitive.

An important feedback was to give the game a backstory to make the actions you make more intuitive and more meaningful. This is something I agree with but is a bit to time consuming for prototyping, so this idea is with us in spirit.