Final iteration
This time I focussed on clarity. In the feedback there where some comments about not knowing what to do, and even though they usually figure it out very fast, I have decided to add an info button for the to game modes so you can always read in more detail how it works.
I have decided against a popup because this is an arcade style game that gets played multiple time in succession by multiple people so this would just be annoying (and personally I don't like pop ups).

I have also added an indicator for when the bottles almost explode so it is clearer what is happening at any given moment.